time to eat first

+++ funscript.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
 mkdir -p /tmp/var/www/
-ln -s /home/akozin/code/webserv/filesforserver /tmp/var/www/filesforserver
-rm /tmp/var/www/filesforserver/funscript.sh
+if [ ! -f /tmp/var/www/filesforserver ]; then
+	ln -s /home/akozin/code/webserv/filesforserver /tmp/var/www/filesforserver
+fi
+rm filesforserver/funscript.sh
